changes to exercise 07 [EX07-PL04-ENT2]

+++ entrega2/ficha4/ex07/seller.c
@@ -15,29 +15,29 @@
 #include <fcntl.h>           /* For O_* constants */
 #include <semaphore.h>
 
+#include "shm_type.h"
+
 // Setting constants
-const char *NAME_QUEUE = "sem_queue";
 const char *NAME_REQUEST = "sem_request_ticket";
 const char *NAME_COLLECT = "sem_collect_ticket";
 const char *NAME_SHM = "sem_shm";
-const char *SHM_NAME = "shm_ticket";
-const int OPEN_TIME = 60;
+const char *NAME_BARRIER1 = "sem_barrier1";
+const char *NAME_BARRIER2 = "sem_barrier2";
+const int FIRST_TICKET = 1000;
 const int NUM_TICKETS = 10;
+
 
 /*
  * PL 4 - Exercise 07 - SELLER
  */
 int main(int argc, char *argv[])
 {
-	sem_t *sem_queue, *sem_request, *sem_collect, *sem_shm;
+	// Print something
+	printf("\nThe THEATER is open!\n");
+	fflush(stdout);
 	
-	// Create a queue semaphore (mutual exclusion)
-	sem_queue = sem_open(NAME_QUEUE, O_CREAT | O_EXCL, S_IRUSR|S_IWUSR, 1);
-	if (sem_queue == SEM_FAILED)
-	{
-		perror("Semaphore failed.\n");
-		exit(EXIT_FAILURE);
-	}
+	sem_t *sem_request, *sem_collect, *sem_shm, *sem_barrier1, *sem_barrier2;
+
 	// Create semaphore to request a ticket (by occurence of events)
 	sem_request = sem_open(NAME_REQUEST, O_CREAT | O_EXCL, S_IRUSR|S_IWUSR, 0);
 	if (sem_request == SEM_FAILED)
@@ -54,14 +54,32 @@ int main(int argc, char *argv[])
 	}
 	// Create a semaphore to access SHM (mutual exclusion)
 	sem_shm = sem_open(NAME_SHM, O_CREAT | O_EXCL, S_IRUSR|S_IWUSR, 1);
-	if (sem_queue == SEM_FAILED)
+	if (sem_shm == SEM_FAILED)
 	{
 		perror("Semaphore failed.\n");
 		exit(EXIT_FAILURE);
 	}
-	
+	// Create a semaphore as a barrier so that each client can check if 
+	// he is next in line (mutual exclusion) 
+	sem_barrier1 = sem_open(NAME_BARRIER1, O_CREAT | O_EXCL, S_IRUSR|S_IWUSR, 1);
+	if (sem_barrier1 == SEM_FAILED)
+	{
+		perror("Semaphore failed.\n");
+		exit(EXIT_FAILURE);
+	}
+	// Create a semaphore as a barrier so that only when all processes that 
+	// have a waiting ticket pass the first barrier. This is to avoid the
+	// possibility of one process controlling the barrier and entering a deadlock (by occurence of events)
+	sem_barrier2 = sem_open(NAME_BARRIER2, O_CREAT | O_EXCL, S_IRUSR|S_IWUSR, 0);
+	if (sem_barrier2 == SEM_FAILED)
+	{
+		perror("Semaphore failed.\n");
+		exit(EXIT_FAILURE);
+	}
+
 	// Shared memory
-	int fd, *ticket_no, data_size = sizeof(ticket_no);
+	shm_type *shm; 
+	int fd, data_size = sizeof(shm_type);
 	// Open shm
 	fd = shm_open(SHM_NAME, O_CREAT|O_EXCL|O_RDWR, S_IRUSR|S_IWUSR);
 	if (fd < 0)
@@ -76,35 +94,32 @@ int main(int argc, char *argv[])
 		exit(EXIT_FAILURE);
 	}
 	// Map shm
-	ticket_no = (int *) mmap(NULL, data_size, PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
-	if (ticket_no == MAP_FAILED)
+	shm = (shm_type *) mmap(NULL, data_size, PROT_READ|PROT_WRITE, MAP_SHARED, fd, 0);
+	if (shm == MAP_FAILED)
 	{
 		perror("Mapping shared memory failed.");
 		exit(EXIT_FAILURE);
 	}
 
 	// Ticket counter
-	int counter = 1;
+	int first_ticket = FIRST_TICKET;
+	int counter = 0;
 	
 	do {
 		sem_wait(sem_request); // Wait for ticket request
+		
 		sem_wait(sem_shm); // Wait for exclusive access to shm
-		
 		// Issue ticket
-		*ticket_no = counter;
-		
+		shm->ticket = first_ticket + counter;
 		sem_post(sem_shm); // Unlock shm
-		sem_post(sem_collect); // Unlock ticket collection 
 		
 		counter++; // next ticket
+		
+		sem_post(sem_collect); // Unlock ticket collection 
+		
 	} while(counter <= NUM_TICKETS);
 	
 	// Close Semaphores
-	if (sem_close(sem_queue) < 0)
-	{
-		perror("SEM close failed.\n");
-		exit(EXIT_FAILURE);
-	}
 	if (sem_close(sem_request) < 0)
 	{
 		perror("SEM close failed.\n");
@@ -120,12 +135,17 @@ int main(int argc, char *argv[])
 		perror("SEM close failed.\n");
 		exit(EXIT_FAILURE);
 	}
-	// Unlink Semaphores
-	if (sem_unlink(NAME_QUEUE) < 0)
+	if (sem_close(sem_barrier1) < 0)
 	{
-		perror("SEM unlink failed.\n");
+		perror("SEM close failed.\n");
 		exit(EXIT_FAILURE);
 	}
+	if (sem_close(sem_barrier2) < 0)
+	{
+		perror("SEM close failed.\n");
+		exit(EXIT_FAILURE);
+	}
+	// Unlink Semaphores
 	if (sem_unlink(NAME_REQUEST) < 0)
 	{
 		perror("SEM unlink failed.\n");
@@ -141,10 +161,20 @@ int main(int argc, char *argv[])
 		perror("SEM unlink failed.\n");
 		exit(EXIT_FAILURE);
 	}
+	if (sem_unlink(NAME_BARRIER1) < 0)
+	{
+		perror("SEM unlink failed.\n");
+		exit(EXIT_FAILURE);
+	}
+	if (sem_unlink(NAME_BARRIER2) < 0)
+	{
+		perror("SEM unlink failed.\n");
+		exit(EXIT_FAILURE);
+	}
 	printf("\nSEMs unlinked.\n");
 	
 	// Unmap & close SHM
-	if (munmap(ticket_no, data_size) < 0)
+	if (munmap(shm, data_size) < 0)
 	{
 		exit(EXIT_FAILURE);
 	}
